+++ bb_proxy/src/packet/v1_9.rs
@@ -13,10 +13,10 @@ pub fn chunk(chunk: ChunkWithPos, ver: ProtocolVersion, conv: &TypeConverter) ->
   );
   let mut chunk_buf = Buffer::new(&mut chunk_data);
   for s in chunk.sections.iter().flatten() {
-    chunk_buf.write_u8(s.data().bpe() as u8);
+    chunk_buf.write_u8(s.data().bpe());
     chunk_buf.write_varint(s.palette().len() as i32);
     for g in s.palette() {
-      chunk_buf.write_varint(conv.block_to_old(*g as u32, ver.block()) as i32);
+      chunk_buf.write_varint(conv.block_to_old(*g, ver.block()) as i32);
     }
     let longs = s.data().old_long_array();
     chunk_buf.write_varint(longs.len() as i32);
